Refact input processing

+++ MatrixLib/3G/Form.hpp
@@ -6,6 +6,7 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 #include "CMain.hpp"
 
@@ -45,7 +46,7 @@ public:
     virtual void MouseMove(int, int) = 0;
     virtual void MouseKey(ButtonStatus, int, int, int) = 0;
 
-    virtual void Keyboard(bool, int) = 0;
+    virtual void Keyboard(bool, uint8_t) = 0;
 
     virtual void SystemEvent(ESysEvent se) = 0;
 };
